Consolidate budget scripts and remove duplicate files
- Removed all _optimized, _yaml, and check_ variations
- Single populate_budget.py now uses YAML config with batch updates
- Personnel tab fixed to show position titles in column B
- Cleaned up directory structure per TDD best practices
